Linking nav graph, adding fragments, removing activities
