+++ androidHackathonApp/src/main/java/com/kcc/kmmhackathon/androidHackathonApp/MainActivity.kt
@@ -1,12 +1,10 @@
-package com.kcc.kmmhackathon.androidHackathonApp.view
+package com.kcc.kmmhackathon.androidHackathonApp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.navigation.findNavController
-import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.kcc.kmmhackathon.androidHackathonApp.R
 
 class MainActivity : AppCompatActivity() {
