+++ androidHackathonApp/src/main/res/menu/bottom_nav_menu.xml
@@ -7,8 +7,12 @@
         android:icon="@drawable/ic_baseline_map_24"
         android:title="@string/map_tab_label" />
 
+    <item android:id="@+id/findMealFragment"
+        android:icon="@drawable/ic_baseline_list_24"
+        android:title="@string/title_switch_to_list"/>
+
     <item
-        android:id="@+id/settingsFragment2"
+        android:id="@+id/settingsFragment"
         android:icon="@drawable/ic_baseline_settings_24"
         android:title="@string/settings_tab_label" />
