+++ androidHackathonApp/src/main/res/values/strings.xml
@@ -15,6 +15,7 @@
     <string name="activity_add_meal_title">AddMealActivity</string>
     <string name="title_email">Email</string>
     <string name="title_password">Password</string>
+    <string name="title_switch_to_list">List View</string>
 
     //Hints
     <string name="hint_email_address">Enter Email address</string>
